+++ CHANGELOG.md
@@ -12,6 +12,8 @@ nav_order: 1
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 
 - Fix `aiven_project`: can't migrate from `account_id` (deprecated) to `parent_id`
+- Add `aiven_organization_user_list` resource
+- Run client-side validation for `aiven_kafka_schema` AVRO type schema
 
 ## [4.27.0] - 2024-10-09
 
@@ -19,7 +21,6 @@ nav_order: 1
 - Removes `receiver_ingesting_remote_write_uri` and `store_uri` Thanos connection info fields
 - Adds `stringtype` to `flink_external_postgresql_user_config` service integration
 - Fix `terraform import` for services with additional disk space or read replica service integration
-- Run client-side validation for `aiven_kafka_schema` AVRO type schema
 
 ## [4.26.0] - 2024-09-25