+++ CHANGELOG.md
@@ -9,10 +9,10 @@ nav_order: 1
 <!-- Always keep the following header in place: -->
 <!-- ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD -->
 
-## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
+## [4.28.0] - 2024-10-21
 
 - Fix `aiven_project`: can't migrate from `account_id` (deprecated) to `parent_id`
-- Add `aiven_organization_user_list` resource
+- Add `aiven_organization_user_list` beta resource
 - Run client-side validation for `aiven_kafka_schema` AVRO type schema
 
 ## [4.27.0] - 2024-10-09